Change data type of centre of mass quantities to float64

+++ property_table.py
@@ -1163,7 +1163,7 @@ class PropertyTable:
         "com": (
             "CentreOfMass",
             3,
-            np.float32,
+            np.float64,
             "Mpc",
             "Centre of mass.",
             "basic",
@@ -1173,7 +1173,7 @@ class PropertyTable:
         "com_gas": (
             "GasCentreOfMass",
             3,
-            np.float32,
+            np.float64,
             "Mpc",
             "Centre of mass of gas.",
             "gas",
@@ -1183,7 +1183,7 @@ class PropertyTable:
         "com_star": (
             "StellarCentreOfMass",
             3,
-            np.float32,
+            np.float64,
             "Mpc",
             "Centre of mass of stars.",
             "star",
